Loading NaNs, interpolation

+++ numflow/cnumflow/cpp/numflow.hpp
@@ -3,26 +3,26 @@
 struct Dataset3D
 {
     int32_t dx, dy, dz;
-    double *ax, *ay, *az;
-    double *data;
+    float *ax, *ay, *az;
+    float *data;
 };
 
 struct DataMatrix
 {
     int32_t rows, columns;
-    double * data;
+    float * data;
 };
 
 struct DataStreamlines
 {
     //positions
-    double * y;
+    float * y;
     int32_t dy;
     //values
-    double * f;
+    float * f;
     int32_t df;
     //times
-    double * t;
+    float * t;
     int32_t dt;
     //lengths
     int32_t * l;
@@ -30,7 +30,7 @@ struct DataStreamlines
 };
 
 //loading functions
-Dataset3D * load_rectilinear_3d(const DataMatrix * mat, double epsilon);
+Dataset3D * load_rectilinear_3d(const DataMatrix * mat, float epsilon);
 DataMatrix * parse_file(const char * filename, const char * sep);
 
 //lod
@@ -38,11 +38,11 @@ Dataset3D * construct_level_3d(const Dataset3D * ds, int32_t x, int32_t y, int32
 
 
 //interpolation
-double * interpolate_3d(const Dataset3D * dataset, const double *points, const int32_t count);
+float * interpolate_3d(const Dataset3D * dataset, const float *points, const int32_t count);
 
 //integration
-DataStreamlines * integrate_3d(const Dataset3D *dataset, double *points,  const int32_t count, 
-                  const double t0, const double tbound);
+DataStreamlines * integrate_3d(const Dataset3D *dataset, float *points,  const int32_t count, 
+                  const float t0, const float tbound);
 
 void delete_dataset_3d(Dataset3D * ds);
 void delete_datamatrix(DataMatrix * dm);
